[#74, #75] changed student registration modal to be used as tutor registration modal, added registration action
