+++ src/containers/guest-home-page/registration-dialog/RegistrationDialog.styles.js
