+++ src/containers/guest-home-page/registration-form/RegistrationForm.styles.js
